+++ src/components/ProgressBar/index.ts
@@ -1,0 +1,3 @@
+export { default as ProgressBar } from './ProgressBar';
+export type { ProgressBarProps } from './ProgressBar';
+export { getProgressBarMetrics } from './utils';